Ignore Nikki's notes and special notes
This patch makes the installer ignore Nikki's notes and special notes when choosing the target note to replace with the doodlebomb. It is achieved with looking at the last 2 bytes of the time field, and pick only those containing `0000`. Nikki's notes end with `XX0E`, while special notes from Nintendo end with `0309`. The only exception is Nikki's welcome note at first startup, which ends in `0000`, but is easy to filter out by its id 0.

All uesr created notes end in `0000`, so user should take care to create a new note that's later than any note they have received. "Later" here means creation time, so if user receives a normal note after creating the target note, as long as the target note is created after the received note, the target note will be replaced.